// src/model/image.ts
/**
 * API BHB Customer Backend
 * API BHB Customer Backend
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Image information
 */
export interface Image { 
    /**
     * Image url
     */
    url?: string;
    /**
     * Image name
     */
    name?: string;
}